% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{codaFromPremium}
\alias{codaFromPremium}
\title{Assemble MCMC samples from \code{link[PReMiuM]{profRegr}} into
    a \code{\link[coda]{mcmc.list}} object with multiple chains.}
\usage{
codaFromPremium(global.parameter, ...)
}
\arguments{
\item{...}{Object(s) of type \code{runInfoObj}, output of
\code{\link[PReMiuM]{profRegr}}.}
}
\value{
A \code{\link[coda]{mcmc.list}}.
}
\description{
Assemble MCMC samples from \code{link[PReMiuM]{profRegr}} into
    a \code{\link[coda]{mcmc.list}} object with multiple chains.
}
