% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getHyperparams}
\alias{getHyperparams}
\title{Get hyperparameters used to fit a PReMiuM model.}
\usage{
getHyperparams(...)
}
\arguments{
\item{...}{Object(s) of type \code{runInfoObj}, output of
\code{\link[PReMiuM]{profRegr}}.}
}
\value{
A matrix with a named column for each model and a named
    row for each hyperparameter.
}
\description{
Get hyperparameters used to fit a PReMiuM model.
}
