% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plotClusterSizes}
\alias{plotClusterSizes}
\title{Plots the cluster sizes}
\usage{
plotClusterSizes(...)
}
\arguments{
\item{...}{Object(s) of type \code{riskProfObj}, output of
\code{\link[PReMiuM]{calcAvgRiskAndProfile}}.}
}
\description{
Plots the cluster sizes
}
