% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plotCovariateProfiles}
\alias{plotCovariateProfiles}
\title{Plot covariate profiles by covariate}
\usage{
plotCovariateProfiles(riskProfObj, whichCovariates = NULL,
  rhoOrder = TRUE, useProfileStar = TRUE, covariate_levels = NULL,
  covariate_labels = NULL)
}
\arguments{
\item{riskProfObj}{Object of type \code{riskProfObj}, output of
\code{\link[PReMiuM]{calcAvgRiskAndProfile}}.}

\item{whichCovariates}{A vector of indices or a vector of strings
corresponding to the covariates that are to be displayed.}

\item{rhoOrder}{Should covariates be ordered in descending order of \code{rho}?
If \code{TRUE}, the default, then indices given in \code{whichCovariates}
will be interpreted as indices of the reordered covariates.
Ignored silently if variable selection was not used.}

\item{useProfileStar}{The definition of the star
profile is given in Liverani, S., Hastie, D. I. and Richardson,
S. (2013) PReMiuM: An R package for Bayesian profile regression.
Ignored silently if variable selection was not used.}

\item{covariate_levels}{Vector of integer values taken by the
(discrete) covariates, specifying the order of the facets.}

\item{covariate_labels}{Vector of strings giving labels for each
level of the covariate, in the same order as \code{covariate_levels}.}
}
\description{
Plots the (discrete) covariate profiles, facetted by covariate,
    similar to the layout in \code{\link[PReMiuM]{plotRiskProfile}}.
}
