% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plotProfilesByCluster}
\alias{plotProfilesByCluster}
\title{Plot covariate profiles by cluster}
\usage{
plotProfilesByCluster(riskProfObj, whichCovariates = NULL,
  rhoOrder = TRUE, useProfileStar = TRUE, covariate_info = list(title
  = "Covariate\\ncategory", levels = NULL, labels = NULL, split = NULL))
}
\arguments{
\item{riskProfObj}{Object of type \code{riskProfObj}, output of
\code{\link[PReMiuM]{calcAvgRiskAndProfile}}.}

\item{whichCovariates}{A vector of indices or a vector of strings
corresponding to the covariates that are to be displayed.}

\item{rhoOrder}{Should covariates be ordered in descending order of \code{rho}?
If \code{TRUE}, the default, then indices given in \code{whichCovariates}
will be interpreted as indices of the reordered covariates.
Ignored silently if variable selection was not used.}

\item{useProfileStar}{The definition of the star
profile is given in Liverani, S., Hastie, D. I. and Richardson,
S. (2013) PReMiuM: An R package for Bayesian profile regression.
Ignored silently if variable selection was not used.}

\item{covariate_info}{Optional list of details about the covariates,
with (some of) the following named elements:
\describe{
    \item{\code{title}:}{String describing the covariate, default
    to "Covariate category".}
    \item{\code{levels}:}{Vector of integer values taken by the
    (discrete) covariates, specifying the order of the facets.}
    \item{\code{labels}:}{Vector of strings giving labels for each
    level of the covariate, in the same order as \code{levels}.}
    \item{\code{split}:}{A character vector specifying the covariate
    grouping for horizontal facetting. The first element appears
    in a subset of covariates to be plotted in a separate horizontal
    facet. The second and third elements, if present, are facet labels
    for covariates that do and do not meet this criterion, respectively,
    e.g. \code{c("_citr", "Citrullinated", "Not citrullinated")}.}
}}
}
\description{
Plots the (discrete) covariate profiles, facetted vertically by cluster and
    (optionally) horizontally by some grouping (\code{split}) of the covariates.
    Covariate categories that are more common within a cluster than they are
    overall are shaded dark; those that are not are shaded light.
}
