% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plotResponse}
\alias{plotResponse}
\title{Plots the (categorical) response profiles, similar to the layout
    of the 'risk' plot in \code{\link[PReMiuM]{plotRiskProfile}}.}
\usage{
plotResponse(riskProfObj, response_levels = NULL, response_labels = NULL)
}
\arguments{
\item{riskProfObj}{Object of type \code{riskProfObj}, output of
\code{\link[PReMiuM]{calcAvgRiskAndProfile}}.}

\item{response_levels}{Vector of integer values taken by the
(categorical) response, specifying the order of the facets.}

\item{response_labels}{Vector of strings giving labels for each
level of the response, in the same order as \code{response_levels}.}
}
\description{
Plots the (categorical) response profiles, similar to the layout
    of the 'risk' plot in \code{\link[PReMiuM]{plotRiskProfile}}.
}
