% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plotRhoDistributions}
\alias{plotRhoDistributions}
\title{Plot distribution of \code{rho} per covariate}
\usage{
plotRhoDistributions(...)
}
\arguments{
\item{...}{Object(s) of type \code{riskProfObj}, output of
\code{\link[PReMiuM]{calcAvgRiskAndProfile}}.}
}
\description{
Plots the mean, median and CI of the variable selection
    parameter \code{rho} for each covariate.
}
