% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plotSimilarityMatrix}
\alias{plotSimilarityMatrix}
\title{Plot similarity matrix}
\usage{
plotSimilarityMatrix(...)
}
\arguments{
\item{...}{Object(s) of type \code{disSimObj}, output of
\code{\link[PReMiuM]{calcDissimilarityMatrix}}.}
}
\description{
Plots the similarity matrix, in the original order and sorted by
    (hierarchical) cluster to illustrate cluster structure.
}
