% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plotVarSelectRho}
\alias{plotVarSelectRho}
\title{Plots the cumulative distribution of the variable selection
    parameter \code{rho}. Can be interpreted as the proportion of
    covariates that would be dropped if only those above a given
    threshold of \code{rho} were included.}
\usage{
plotVarSelectRho(...)
}
\arguments{
\item{...}{Object(s) of type \code{riskProfObj}, output of
\code{\link[PReMiuM]{calcAvgRiskAndProfile}}.}
}
\description{
Plots the cumulative distribution of the variable selection
    parameter \code{rho}. Can be interpreted as the proportion of
    covariates that would be dropped if only those above a given
    threshold of \code{rho} were included.
}
