% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{renderPremiumReport}
\alias{renderPremiumReport}
\title{Render PReMiuM report}
\usage{
renderPremiumReport(..., filename = "premium_report.html",
  rmd.template = NULL)
}
\arguments{
\item{...}{Object(s) of type \code{runInfoObj}, output of
\code{\link[PReMiuM]{profRegr}}.}

\item{filename}{File name (.html) to save markdown report.}

\item{rmd.template}{An alternate Rmd template to populate using
the models in \code{...}.}
}
\value{
An HTML file written to \code{filename}.
}
\description{
Render a standardised HTML report on PReMiuM model(s) using
    markdown, with covariate/response profiles and MCMC diagnostics.
}
