% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{summariseVarSelectRhoFromRiskProfObj}
\alias{summariseVarSelectRhoFromRiskProfObj}
\title{Summarise \code{rho}}
\usage{
summariseVarSelectRhoFromRiskProfObj(riskProfObj)
}
\arguments{
\item{riskProfObj}{Object of type \code{riskProfObj}, output of
\code{\link[PReMiuM]{calcAvgRiskAndProfile}}.}
}
\value{
A \code{\link[tibble]{tibble}} containing the mean, median,
    lower (0.05) and upper (0.95) quantiles for each covariate,
    and a ranking from the covariate with the highest value of
    \code{rho} to the lowest.
}
\description{
Summarise the posterior distribution of the variable selection
    parameter \code{rho}. Modified version of
    \code{\link[PReMiuM]{summariseVarSelectRho}}, allowing plotting
    functions to access \code{rho} summaries.
}
