% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tabulateCovariateProfiles}
\alias{tabulateCovariateProfiles}
\title{Make table of covariate profiles with info for plotting}
\usage{
tabulateCovariateProfiles(riskProfObj, whichCovariates = NULL,
  rhoOrder = TRUE, useProfileStar = TRUE, rho_file = NULL)
}
\arguments{
\item{riskProfObj}{Object of type \code{riskProfObj}, output of
\code{\link[PReMiuM]{calcAvgRiskAndProfile}}.}

\item{whichCovariates}{A vector of indices or a vector of strings
corresponding to the covariates that are to be displayed.}

\item{rhoOrder}{Should covariates be ordered in descending order of \code{rho}?
If \code{TRUE}, the default, then indices given in \code{whichCovariates}
will be interpreted as indices of the reordered covariates.
Ignored silently if variable selection was not used.}

\item{useProfileStar}{The definition of the star
profile is given in Liverani, S., Hastie, D. I. and Richardson,
S. (2013) PReMiuM: An R package for Bayesian profile regression.
Ignored silently if variable selection was not used.}

\item{rho_file}{Path to the rho.txt file written by \code{\link[PReMiuM]{profRegr}}.
If NULL (the default), path is taken from the \code{ClusObjRunInfoObj} element
of \code{riskProfObj}.}
}
\description{
Used by \code{\link{plotProfilesByCluster}} and \code{\link{plotCovariateProfiles}}
}
