% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tabulateVarSelectRho}
\alias{tabulateVarSelectRho}
\title{Summarise \code{rho}}
\usage{
tabulateVarSelectRho(riskProfObj, rho_file = NULL)
}
\arguments{
\item{riskProfObj}{Object of type \code{riskProfObj}, output of
\code{\link[PReMiuM]{calcAvgRiskAndProfile}}.}

\item{rho_file}{Path to the rho.txt file written by \code{\link[PReMiuM]{profRegr}}.
If NULL (the default), path is taken from the \code{ClusObjRunInfoObj} element
of \code{riskProfObj}.}
}
\value{
A \code{\link[tibble]{tibble}} containing the mean, median,
    lower (0.05) and upper (0.95) quantiles for each covariate,
    and a ranking from the covariate with the highest value of
    \code{rho} to the lowest.
}
\description{
Summarise the posterior distribution of the variable selection
    parameter \code{rho}. Modified version of
    \code{\link[PReMiuM]{summariseVarSelectRho}}, allowing plotting
    functions to access \code{rho} summaries.
}
