% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plotProfilesByCluster}
\alias{plotProfilesByCluster}
\title{Plot covariate profiles by cluster}
\usage{
plotProfilesByCluster(riskProfObj, whichCovariates = NULL,
  rhoMinimum = NULL, useProfileStar = TRUE, covariate_levels = NULL,
  covariate_labels = NULL, covariate_split = NULL)
}
\arguments{
\item{riskProfObj}{Object of type \code{riskProfObj}, output of
\code{\link[PReMiuM]{calcAvgRiskAndProfile}}.}

\item{whichCovariates}{A vector of indices or a vector of strings
corresponding to the covariates that are to be displayed. If
a single integer, this is intepreted as the 'top N' covariates,
as ranked by their value of \code{rho} (requires the model to
have been fitted with variable selection).}

\item{rhoMinimum}{As an alternative to \code{whichCovariates}, plot
plot all covariates whose value of \code{rho} is greater than
some threshold (requires the model to have been fitted with
variable selection).}

\item{useProfileStar}{o be set equal to TRUE only if a variable
selection procedure has been run. The definition of the star
profile is given in Liverani, S., Hastie, D. I. and Richardson,
S. (2013) PReMiuM: An R package for Bayesian profile regression.}

\item{covariate_levels}{Vector of integer values taken by the
(discrete) covariates, specifying the order of the facets.}

\item{covariate_labels}{Vector of strings giving labels for each
level of the covariate, in the same order as \code{covariate_levels}.}

\item{covariate_split}{A character vector. The first element appears
in a subsets of covariates to be plotted in a separate horizontal
facet. The second and third elements, if present, are facet labels
for covariates that do and do not have meet this criterion, respectively,
e.g. \code{c("_citr", "Citrullinated proteins", "Not citrullinated")}.}
}
\description{
Plots the (discrete) covariate profiles, facetted by cluster.
}
